Move OnboardingWrapper to its own file
